Encoder wraparound on load/save page

+++ shruthi/editor.cc
@@ -625,14 +625,14 @@ const prog_uint8_t default_name[] PROGMEM = "user    ";
 /* static */
 void Editor::set_edited_item_number(int16_t value) {
   if (load_save_target_ & LOAD_SAVE_TARGET_PATCH) {
-    if (value >= Storage::size<Patch>()) {
+    if (value >= static_cast<int16_t>(Storage::size<Patch>())) {
       value = 0;
     } else if (value < 0) {
       value = Storage::size<Patch>() - 1;
     }
     current_patch_number_ = value;
   } else {
-    if (value >= Storage::size<SequencerSettings>()) {
+    if (value >= static_cast<int16_t>(Storage::size<SequencerSettings>())) {
       value = 0;
     } else if (value < 0) {
       value = Storage::size<SequencerSettings>() - 1;
